pyproject.toml
[project]
name = "rl-overview"
version = "0.1.0"
description = "Reinforcement Learning Overview — didactic, code-as-textbook project"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
authors = [{ name = "rl-overview" }]

dependencies = [
  "hydra-core>=1.3",
  "numpy>=1.24",
  "tqdm>=4.66",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
]
torch = [
  "torch>=2.2",
]
gym = [
  "gymnasium>=0.29",
]

[project.urls]
Homepage = "https://example.com/rl-overview"

[tool.uv]
package = true

[tool.uv.sources]

[tool.uv.workspace]
members = []

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
